+++ src/helpers/log_helpers.ts
@@ -72,7 +72,7 @@ export async function lockExecution(
 };
 
 
-// Log only if DEBUG env var is set.
+// Log only if FEDIALGO_DEBUG env var is set to "true"
 // Assumes if there's multiple args and the 2nd one is a string the 1st one is a prefix.
 export function traceLog(msg: string, ...args: any[]): void {
     if (!ENABLE_TRACE_LOG) return;